refactor: split the GitHub exporter into scripts/gh_export/, and label review threads resolved
Ported from vzakharov/agent-project-boilerplate@71c6201.

The exporter was a single 616-line script — well past CLAUDE.md's
~450-line guidance and with six unrelated concerns in it. It becomes a
155-line entrypoint over one module per concern: argument parsing, the
REST/GraphQL client, attachment download, and a renderer each for the
header and comments, the review threads, and the timeline. `scripts/` is
already on `sys.path[0]` for a script run as `python3 scripts/<name>.py`,
so `gh_export` resolves as a PEP 420 namespace package exactly as
`lib.github` does — no `__init__.py`, no path manipulation, still stdlib
only.

A PR export now states each review thread's resolved / unresolved /
resolution-unknown state in its header. Resolution is the one fact no
REST comment payload carries, so it comes from GraphQL, keyed by every
comment id in the thread rather than the root: the export rebuilds
threads from REST `in_reply_to_id` chains, which root on a different
comment when the parent falls off the page. A thread the query didn't
return reads as `resolution unknown` rather than silently as open,
because a reader acts on that word.

Verified against PRs #25 and #26: the labels the export writes match the
GraphQL `isResolved` truth thread for thread.
